+++ e2e/fixtures/cache/src/pages/index.ts
@@ -1,0 +1,24 @@
+import { eco } from '@ecopages/core';
+import { BaseLayout } from '../layouts/base-layout';
+
+export default eco.page({
+	metadata: () => ({
+		title: 'Home page',
+		description: 'This is the homepage of the website',
+		image: 'public/assets/images/default-og.png',
+		keywords: ['typescript', 'framework', 'static'],
+	}),
+	dependencies: {
+		stylesheets: ['./index.css'],
+		components: [BaseLayout],
+	},
+	render: ({ params, query }) =>
+		`${BaseLayout({
+			class: 'main-content',
+			children: `<h1 class="main-title">Home Page</h1>
+				<a href="/dynamic/a">Dynamic</a>
+				<a href="/dynamic/b?q=query">With Query</a>
+				<p>${JSON.stringify(query || [])}</p>
+				<p>${JSON.stringify(params || [])}</p>`,
+		})}`,
+});